make deliv check fixture public
